[WIP] on importing terms
The problem here was getting the CSV to escape properly
and play nicely with the internal field separator in Bash.

I ended up doing the import manually for the terms since there
weren't many to do in this job (it was mostly CPTs). The lesson
learned was to use PHP because, as one can see, I had to make
a PHP script to serialize the date here anyway. More experience
in Bash would have helped here, of course.

In the end, the most effective scripts were those written in PHP.
No surprise since I knew it better and WP is native to PHP.